requirements.txt
matplotlib==3.5.1
numpy==1.21.5
pandas==1.3.5
Pillow==9.1.1
plotly==5.7.0
seaborn==0.11.2
streamlit==1.9.0
